Removed extra logadd function that was not needed, instead it is just the code of Grapher.cc that had to be fixed to compile with -float

+++ plearn_learners/misc/Grapher.cc
@@ -442,7 +442,7 @@ void DX_create_grid_outputs_file(const string& filename, PP<PLearner> learner, V
             if(is_equal(logsum, -FLT_MAX))
                 logsum = output[0];
             else 
-                logsum = logadd(logsum, output[0]);
+                logsum = logadd(logsum, double(output[0]));
             pb.update(n++);
         }
     }
